fix format warnings for extreme debugging
git-svn-id: http://subversion.ffado.org/ffado/branches/libffado-2.0@1722 2be59082-3212-0410-8809-b0798e1608f0

+++ src/libieee1394/IsoHandler.cpp
@@ -396,9 +396,13 @@ enum raw1394_iso_disposition IsoHandler::putPacket(
 
     if((pkt_ctr & ~0x0FFFL) != pkt_ctr_ref) {
         debugWarning("reconstructed CTR counter discrepancy\n");
-        debugWarning(" ingredients: %X, %X, %X, %X, %X, %d, %ld, %ld, %ld\n",
-                       cycle, pkt_ctr_ref, pkt_ctr, now, m_last_now, now_secs_ref, CYCLE_TIMER_GET_SECS(now), CYCLE_TIMER_GET_SECS(m_last_now), tmp);
-        debugWarning(" diffcy = %ld \n", diff_cycles);
+        debugWarning(" ingredients: %X, %X, %X, %X, %X, %d, %ld, %ld, %"PRId64"\n",
+                     cycle, pkt_ctr_ref, pkt_ctr, 
+		     now, m_last_now, now_secs_ref, 
+		     (long int)CYCLE_TIMER_GET_SECS(now),
+		     (long int)CYCLE_TIMER_GET_SECS(m_last_now),
+		     tmp);
+        debugWarning(" diffcy = %"PRId64" \n", diff_cycles);
     }
     #endif
     m_last_packet_handled_at = pkt_ctr;
@@ -489,9 +493,13 @@ IsoHandler::getPacket(unsigned char *data, unsigned int *length,
 
         if(((pkt_ctr & ~0x0FFFL) != pkt_ctr_ref) && (m_packets > m_buf_packets)) {
             debugWarning("reconstructed CTR counter discrepancy\n");
-            debugWarning(" ingredients: %X, %X, %X, %X, %X, %d, %ld, %ld, %ld\n",
-                        cycle, pkt_ctr_ref, pkt_ctr, now, m_last_now, now_secs_ref, CYCLE_TIMER_GET_SECS(now), CYCLE_TIMER_GET_SECS(m_last_now), tmp);
-            debugWarning(" diffcy = %ld \n", diff_cycles);
+            debugWarning(" ingredients: %X, %X, %X, %X, %X, %d, %ld, %ld, %"PRId64"\n",
+                        cycle, pkt_ctr_ref, pkt_ctr,
+			 now, m_last_now, now_secs_ref, 
+			 (long int)CYCLE_TIMER_GET_SECS(now),
+			 (long int)CYCLE_TIMER_GET_SECS(m_last_now), 
+			 tmp);
+            debugWarning(" diffcy = %"PRId64" \n", diff_cycles);
         }
         #endif
     }
