+++ src/libstreaming/generic/StreamProcessor.cpp
@@ -377,7 +377,7 @@ StreamProcessor::putPacket(unsigned char *data, unsigned int length,
         #ifdef DEBUG
         if (m_last_timestamp > 0 && m_last_timestamp2 > 0) {
             int64_t tsp_diff = diffTicks(m_last_timestamp, m_last_timestamp2);
-            debugOutputExtreme(DEBUG_LEVEL_VERBOSE, "TSP diff: %ld\n", tsp_diff);
+            debugOutputExtreme(DEBUG_LEVEL_VERBOSE, "TSP diff: %"PRId64"\n", tsp_diff);
             double tsp_diff_d = tsp_diff;
             double fs_syt = 1.0/tsp_diff_d;
             fs_syt *= (double)getNominalFramesPerPacket() * (double)TICKS_PER_USEC * 1e6;
@@ -398,8 +398,8 @@ StreamProcessor::putPacket(unsigned char *data, unsigned int length,
                 // so 50 ticks = 10%, which is a rather large jitter value.
                 if(diff-ticks_per_packet > 50 || diff-ticks_per_packet < -50) {
                     debugOutput(DEBUG_LEVEL_VERBOSE,
-                                "cy %04lu rather large TSP difference TS=%011lu => TS=%011lu (%d, nom %d)\n",
-                                CYCLE_TIMER_GET_CYCLES(pkt_ctr), m_last_timestamp2,
+                                "cy %04d rather large TSP difference TS=%011"PRIu64" => TS=%011"PRIu64" (%d, nom %d)\n",
+                                (int)CYCLE_TIMER_GET_CYCLES(pkt_ctr), m_last_timestamp2,
                                 m_last_timestamp, diff, ticks_per_packet);
                     // !!!HACK!!! FIXME: this is the result of a failure in wrapping/unwrapping somewhere
                     // it's definitely a bug.
@@ -409,8 +409,8 @@ StreamProcessor::putPacket(unsigned char *data, unsigned int length,
                     last_timestamp_fixed = addTicks(m_last_timestamp, TICKS_PER_SECOND);
                     diff = diffTicks(last_timestamp_fixed, m_last_timestamp2);
                     if(diff-ticks_per_packet < 50 && diff-ticks_per_packet > -50) {
-                        debugWarning("cy %04lu rather large TSP difference TS=%011lu => TS=%011lu (%d, nom %d)\n",
-                                    CYCLE_TIMER_GET_CYCLES(pkt_ctr), m_last_timestamp2,
+                        debugWarning("cy %04d rather large TSP difference TS=%011"PRIu64" => TS=%011"PRIu64" (%d, nom %d)\n",
+                                    (int)CYCLE_TIMER_GET_CYCLES(pkt_ctr), m_last_timestamp2,
                                     m_last_timestamp, diff, ticks_per_packet);
                         debugWarning("HACK: fixed by adding one second of ticks. This is a bug being run-time fixed.\n");
                         m_last_timestamp = last_timestamp_fixed;
@@ -419,8 +419,8 @@ StreamProcessor::putPacket(unsigned char *data, unsigned int length,
                         last_timestamp_fixed = substractTicks(m_last_timestamp, TICKS_PER_SECOND);
                         diff = diffTicks(last_timestamp_fixed, m_last_timestamp2);
                         if(diff-ticks_per_packet < 50 && diff-ticks_per_packet > -50) {
-                            debugWarning("cy %04lu rather large TSP difference TS=%011lu => TS=%011lu (%d, nom %d)\n",
-                                        CYCLE_TIMER_GET_CYCLES(pkt_ctr), m_last_timestamp2,
+                            debugWarning("cy %04d rather large TSP difference TS=%011"PRIu64" => TS=%011"PRIu64" (%d, nom %d)\n",
+                                        (int)CYCLE_TIMER_GET_CYCLES(pkt_ctr), m_last_timestamp2,
                                         m_last_timestamp, diff, ticks_per_packet);
                             debugWarning("HACK: fixed by subtracing one second of ticks. This is a bug being run-time fixed.\n");
                             m_last_timestamp = last_timestamp_fixed;
@@ -428,16 +428,16 @@ StreamProcessor::putPacket(unsigned char *data, unsigned int length,
                     }
                 }
                 debugOutputExtreme(DEBUG_LEVEL_VERY_VERBOSE,
-                                "%04u %011lu %011lu %d %d\n",
-                                CYCLE_TIMER_GET_CYCLES(pkt_ctr),
+                                "%04u %011"PRIu64" %011"PRIu64" %d %d\n",
+                                (int)CYCLE_TIMER_GET_CYCLES(pkt_ctr),
                                 m_last_timestamp2, m_last_timestamp, 
                                 diff, ticks_per_packet);
         }
         #endif
 
         debugOutputExtreme(DEBUG_LEVEL_VERY_VERBOSE,
-                          "RECV: CY=%04u TS=%011lu\n",
-                          CYCLE_TIMER_GET_CYCLES(pkt_ctr),
+                          "RECV: CY=%04u TS=%011"PRIu64"\n",
+                          (int)CYCLE_TIMER_GET_CYCLES(pkt_ctr),
                           m_last_timestamp);
 
         if(m_correct_last_timestamp) {
@@ -605,8 +605,8 @@ StreamProcessor::getPacket(unsigned char *data, unsigned int *length,
         enum eChildReturnValue result = generateSilentPacketHeader(data, length, tag, sy, pkt_ctr);
         if (result == eCRV_Packet) {
             debugOutputExtreme(DEBUG_LEVEL_VERY_VERBOSE,
-                               "XMIT SILENT: CY=%04u TS=%011lu\n",
-                               CYCLE_TIMER_GET_CYCLES(pkt_ctr), m_last_timestamp);
+                               "XMIT SILENT: CY=%04u TS=%011"PRIu64"\n",
+                               (int)CYCLE_TIMER_GET_CYCLES(pkt_ctr), m_last_timestamp);
 
             // assumed not to xrun
             generateSilentPacketData(data, length);
@@ -656,8 +656,8 @@ StreamProcessor::getPacket(unsigned char *data, unsigned int *length,
         enum eChildReturnValue result = generatePacketHeader(data, length, tag, sy, pkt_ctr);
         if (result == eCRV_Packet || result == eCRV_Defer) {
             debugOutputExtreme(DEBUG_LEVEL_VERBOSE,
-                               "XMIT: CY=%04u TS=%011lu\n",
-                               CYCLE_TIMER_GET_CYCLES(pkt_ctr), m_last_timestamp);
+                               "XMIT: CY=%04u TS=%011"PRIu64"\n",
+                               (int)CYCLE_TIMER_GET_CYCLES(pkt_ctr), m_last_timestamp);
 
             // valid packet timestamp
             m_last_timestamp2 = prev_timestamp;
@@ -693,7 +693,7 @@ StreamProcessor::getPacket(unsigned char *data, unsigned int *length,
             #ifdef DEBUG
             if (m_last_timestamp > 0 && m_last_timestamp2 > 0) {
                 int64_t tsp_diff = diffTicks(m_last_timestamp, m_last_timestamp2);
-                debugOutputExtreme(DEBUG_LEVEL_VERBOSE, "TSP diff: %ld\n", tsp_diff);
+                debugOutputExtreme(DEBUG_LEVEL_VERBOSE, "TSP diff: %"PRId64"\n", tsp_diff);
                 double tsp_diff_d = tsp_diff;
                 double fs_syt = 1.0/tsp_diff_d;
                 fs_syt *= (double)getNominalFramesPerPacket() * (double)TICKS_PER_USEC * 1e6;
@@ -702,7 +702,7 @@ StreamProcessor::getPacket(unsigned char *data, unsigned int *length,
                 double fs_diff_norm = fs_diff/fs_nom;
                 debugOutputExtreme(DEBUG_LEVEL_VERBOSE, "Nom fs: %12f, Instantanous fs: %12f, diff: %12f (%12f)\n",
                            fs_nom, fs_syt, fs_diff, fs_diff_norm);
-//                 debugOutput(DEBUG_LEVEL_VERBOSE, "Diff fs: %12f, m_last_timestamp: %011lu, m_last_timestamp2: %011lu\n",
+//                 debugOutput(DEBUG_LEVEL_VERBOSE, "Diff fs: %12f, m_last_timestamp: %011"PRIu64", m_last_timestamp2: %011"PRIu64"\n",
 //                            fs_diff, m_last_timestamp, m_last_timestamp2);
                 if (fs_diff_norm > 0.01 || fs_diff_norm < -0.01) {
                     debugWarning( "Instantanous samplerate more than 1%% off nominal. [Nom fs: %12f, Instantanous fs: %12f, diff: %12f (%12f)]\n",
@@ -715,13 +715,13 @@ StreamProcessor::getPacket(unsigned char *data, unsigned int *length,
                 // so 50 ticks = 10%, which is a rather large jitter value.
                 if(diff-ticks_per_packet > 50 || diff-ticks_per_packet < -50) {
                     debugOutput(DEBUG_LEVEL_VERBOSE,
-                                "cy %04ld, rather large TSP difference TS=%011lu => TS=%011lu (%d, nom %d)\n",
-                                CYCLE_TIMER_GET_CYCLES(pkt_ctr), m_last_timestamp2,
+                                "cy %04d, rather large TSP difference TS=%011"PRIu64" => TS=%011"PRIu64" (%d, nom %d)\n",
+                                (int)CYCLE_TIMER_GET_CYCLES(pkt_ctr), m_last_timestamp2,
                                 m_last_timestamp, diff, ticks_per_packet);
                 }
                 debugOutputExtreme(DEBUG_LEVEL_VERY_VERBOSE,
-                                "%04ld %011lu %011lu %d %d\n",
-                                CYCLE_TIMER_GET_CYCLES(pkt_ctr), m_last_timestamp2,
+                                "%04d %011"PRIu64" %011"PRIu64" %d %d\n",
+                                (int)CYCLE_TIMER_GET_CYCLES(pkt_ctr), m_last_timestamp2,
                                 m_last_timestamp, diff, ticks_per_packet);
             }
             #endif
@@ -756,7 +756,7 @@ StreamProcessor::getPacket(unsigned char *data, unsigned int *length,
             }
             goto send_empty_packet;
         } else if (result == eCRV_Again) {
-            debugOutput(DEBUG_LEVEL_VERY_VERBOSE, "have to retry cycle %ld\n", CYCLE_TIMER_GET_CYCLES(pkt_ctr));
+            debugOutput(DEBUG_LEVEL_VERY_VERBOSE, "have to retry cycle %d\n", (int)CYCLE_TIMER_GET_CYCLES(pkt_ctr));
             if(m_state != m_next_state) {
                 debugOutput(DEBUG_LEVEL_VERBOSE, "Should update state from %s to %s\n",
                                                 ePSToString(m_state), ePSToString(m_next_state));
@@ -791,7 +791,7 @@ send_empty_packet:
     }
     debugOutputExtreme(DEBUG_LEVEL_VERBOSE,
                        "XMIT EMPTY: CY=%04u\n",
-                       CYCLE_TIMER_GET_CYCLES(pkt_ctr));
+                       (int)CYCLE_TIMER_GET_CYCLES(pkt_ctr));
 
     generateEmptyPacketHeader(data, length, tag, sy, pkt_ctr);
     generateEmptyPacketData(data, length);
@@ -808,7 +808,7 @@ send_empty_packet:
 bool StreamProcessor::getFrames(unsigned int nbframes, int64_t ts) {
     bool result;
     debugOutputExtreme( DEBUG_LEVEL_VERBOSE,
-                        "(%p, %s) getFrames(%d, %11lu)\n",
+                        "(%p, %s) getFrames(%d, %11"PRIu64")\n",
                         this, getTypeString(), nbframes, ts);
     assert( getType() == ePT_Receive );
     if(isDryRunning()) result = getFramesDry(nbframes, ts);
@@ -839,15 +839,15 @@ bool StreamProcessor::getFramesWet(unsigned int nbframes, int64_t ts) {
     lag_ticks = diffTicks(ts, ts_expected);
     lag_frames = (((float)lag_ticks) / srate);
     debugOutputExtreme(DEBUG_LEVEL_VERY_VERBOSE,
-                       "stream (%p): drifts %6d ticks = %10.5f frames (rate=%10.5f), %ld, %lu, %d\n",
+                       "stream (%p): drifts %6d ticks = %10.5f frames (rate=%10.5f), %"PRId64", %"PRIu64", %d\n",
                        this, lag_ticks, lag_frames, srate, ts, ts_expected, fc);
     if (lag_frames >= 1.0) {
         // the stream lags
-        debugOutput(DEBUG_LEVEL_VERBOSE, "stream (%p): lags  with %6d ticks = %10.5f frames (rate=%10.5f), %ld, %lu, %d\n",
+        debugOutput(DEBUG_LEVEL_VERBOSE, "stream (%p): lags  with %6d ticks = %10.5f frames (rate=%10.5f), %"PRId64", %"PRIu64", %d\n",
                       this, lag_ticks, lag_frames, srate, ts, ts_expected, fc);
     } else if (lag_frames <= -1.0) {
         // the stream leads
-        debugOutput(DEBUG_LEVEL_VERBOSE, "stream (%p): leads with %6d ticks = %10.5f frames (rate=%10.5f), %ld, %lu, %d\n",
+        debugOutput(DEBUG_LEVEL_VERBOSE, "stream (%p): leads with %6d ticks = %10.5f frames (rate=%10.5f), %"PRId64", %"PRIu64", %d\n",
                       this, lag_ticks, lag_frames, srate, ts, ts_expected, fc);
     }
 #endif
@@ -861,7 +861,7 @@ bool StreamProcessor::getFramesWet(unsigned int nbframes, int64_t ts) {
 bool StreamProcessor::getFramesDry(unsigned int nbframes, int64_t ts)
 {
     debugOutputExtreme(DEBUG_LEVEL_VERY_VERBOSE,
-                       "stream (%p): dry run %d frames (@ ts=%ld)\n",
+                       "stream (%p): dry run %d frames (@ ts=%"PRId64")\n",
                        this, nbframes, ts);
     // dry run on this side means that we put silence in all enabled ports
     // since there is do data put into the ringbuffer in the dry-running state
@@ -872,7 +872,7 @@ bool
 StreamProcessor::dropFrames(unsigned int nbframes, int64_t ts)
 {
     bool result;
-    debugOutput(DEBUG_LEVEL_VERY_VERBOSE, "StreamProcessor::dropFrames(%d, %ld)\n", nbframes, ts);
+    debugOutput(DEBUG_LEVEL_VERY_VERBOSE, "StreamProcessor::dropFrames(%d, %"PRId64")\n", nbframes, ts);
     result = m_data_buffer->dropFrames(nbframes);
     SIGNAL_ACTIVITY_ISO_RECV;
     return result;
@@ -882,7 +882,7 @@ bool StreamProcessor::putFrames(unsigned int nbframes, int64_t ts)
 {
     bool result;
     debugOutputExtreme( DEBUG_LEVEL_VERBOSE,
-                        "(%p, %s) putFrames(%d, %11lu)\n",
+                        "(%p, %s) putFrames(%d, %11"PRIu64")\n",
                         this, getTypeString(), nbframes, ts);
     assert( getType() == ePT_Transmit );
     if(isDryRunning()) result = putFramesDry(nbframes, ts);
@@ -895,12 +895,12 @@ bool
 StreamProcessor::putFramesWet(unsigned int nbframes, int64_t ts)
 {
     debugOutputExtreme(DEBUG_LEVEL_ULTRA_VERBOSE,
-                       "StreamProcessor::putFramesWet(%d, %lu)\n",
+                       "StreamProcessor::putFramesWet(%d, %"PRIu64")\n",
                        nbframes, ts);
     // transfer the data
     m_data_buffer->blockProcessWriteFrames(nbframes, ts);
     debugOutputExtreme(DEBUG_LEVEL_ULTRA_VERBOSE,
-                       " New timestamp: %lu\n", ts);
+                       " New timestamp: %"PRIu64"\n", ts);
     return true; // FIXME: what about failure?
 }
 
@@ -908,7 +908,7 @@ bool
 StreamProcessor::putFramesDry(unsigned int nbframes, int64_t ts)
 {
     debugOutputExtreme(DEBUG_LEVEL_ULTRA_VERBOSE,
-                       "StreamProcessor::putFramesDry(%d, %lu)\n",
+                       "StreamProcessor::putFramesDry(%d, %"PRIu64")\n",
                        nbframes, ts);
     // do nothing
     return true;
@@ -918,7 +918,7 @@ bool
 StreamProcessor::putSilenceFrames(unsigned int nbframes, int64_t ts)
 {
     debugOutput(DEBUG_LEVEL_VERY_VERBOSE,
-                       "StreamProcessor::putSilenceFrames(%d, %lu)\n",
+                       "StreamProcessor::putSilenceFrames(%d, %"PRIu64")\n",
                        nbframes, ts);
 
     size_t bytes_per_frame = getEventSize() * getEventsPerFrame();
@@ -1069,7 +1069,7 @@ bool StreamProcessor::prepare()
 
     // make the scratch buffer one period of frames long
     m_scratch_buffer_size_bytes = m_StreamProcessorManager.getPeriodSize() * getEventsPerFrame() * getEventSize();
-    debugOutput( DEBUG_LEVEL_VERBOSE, " Allocate scratch buffer of %ld quadlets\n", m_scratch_buffer_size_bytes);
+    debugOutput( DEBUG_LEVEL_VERBOSE, " Allocate scratch buffer of %zd quadlets\n", m_scratch_buffer_size_bytes);
     if(m_scratch_buffer) delete[] m_scratch_buffer;
     m_scratch_buffer = new byte_t[m_scratch_buffer_size_bytes];
     if(m_scratch_buffer == NULL) {
@@ -1158,12 +1158,12 @@ bool StreamProcessor::scheduleStartDryRunning(int64_t t) {
     debugOutput(DEBUG_LEVEL_VERBOSE,"for %s SP (%p)\n", ePTToString(getType()), this);
     #ifdef DEBUG
     uint64_t now = m_1394service.getCycleTimerTicks();
-    debugOutput(DEBUG_LEVEL_VERBOSE,"  Now                   : %011lu (%03us %04uc %04ut)\n",
+    debugOutput(DEBUG_LEVEL_VERBOSE,"  Now                   : %011"PRIu64" (%03us %04uc %04ut)\n",
                           now,
                           (unsigned int)TICKS_TO_SECS(now),
                           (unsigned int)TICKS_TO_CYCLES(now),
                           (unsigned int)TICKS_TO_OFFSET(now));
-    debugOutput(DEBUG_LEVEL_VERBOSE,"  Start at              : %011lu (%03us %04uc %04ut)\n",
+    debugOutput(DEBUG_LEVEL_VERBOSE,"  Start at              : %011"PRIu64" (%03us %04uc %04ut)\n",
                           tx,
                           (unsigned int)TICKS_TO_SECS(tx),
                           (unsigned int)TICKS_TO_CYCLES(tx),
@@ -1204,12 +1204,12 @@ bool StreamProcessor::scheduleStartRunning(int64_t t) {
     debugOutput(DEBUG_LEVEL_VERBOSE,"for %s SP (%p)\n", ePTToString(getType()), this);
     #ifdef DEBUG
     uint64_t now = m_1394service.getCycleTimerTicks();
-    debugOutput(DEBUG_LEVEL_VERBOSE,"  Now                   : %011lu (%03us %04uc %04ut)\n",
+    debugOutput(DEBUG_LEVEL_VERBOSE,"  Now                   : %011"PRIu64" (%03us %04uc %04ut)\n",
                           now,
                           (unsigned int)TICKS_TO_SECS(now),
                           (unsigned int)TICKS_TO_CYCLES(now),
                           (unsigned int)TICKS_TO_OFFSET(now));
-    debugOutput(DEBUG_LEVEL_VERBOSE,"  Start at              : %011lu (%03us %04uc %04ut)\n",
+    debugOutput(DEBUG_LEVEL_VERBOSE,"  Start at              : %011"PRIu64" (%03us %04uc %04ut)\n",
                           tx,
                           (unsigned int)TICKS_TO_SECS(tx),
                           (unsigned int)TICKS_TO_CYCLES(tx),
@@ -1228,12 +1228,12 @@ bool StreamProcessor::scheduleStopDryRunning(int64_t t) {
     debugOutput(DEBUG_LEVEL_VERBOSE,"for %s SP (%p)\n", ePTToString(getType()), this);
     #ifdef DEBUG
     uint64_t now = m_1394service.getCycleTimerTicks();
-    debugOutput(DEBUG_LEVEL_VERBOSE,"  Now                   : %011lu (%03us %04uc %04ut)\n",
+    debugOutput(DEBUG_LEVEL_VERBOSE,"  Now                   : %011"PRIu64" (%03us %04uc %04ut)\n",
                           now,
                           (unsigned int)TICKS_TO_SECS(now),
                           (unsigned int)TICKS_TO_CYCLES(now),
                           (unsigned int)TICKS_TO_OFFSET(now));
-    debugOutput(DEBUG_LEVEL_VERBOSE,"  Stop at              : %011lu (%03us %04uc %04ut)\n",
+    debugOutput(DEBUG_LEVEL_VERBOSE,"  Stop at              : %011"PRIu64" (%03us %04uc %04ut)\n",
                           tx,
                           (unsigned int)TICKS_TO_SECS(tx),
                           (unsigned int)TICKS_TO_CYCLES(tx),
@@ -1253,12 +1253,12 @@ bool StreamProcessor::scheduleStopRunning(int64_t t) {
     debugOutput(DEBUG_LEVEL_VERBOSE,"for %s SP (%p)\n", ePTToString(getType()), this);
     #ifdef DEBUG
     uint64_t now = m_1394service.getCycleTimerTicks();
-    debugOutput(DEBUG_LEVEL_VERBOSE,"  Now                   : %011lu (%03us %04uc %04ut)\n",
+    debugOutput(DEBUG_LEVEL_VERBOSE,"  Now                   : %011"PRIu64" (%03us %04uc %04ut)\n",
                           now,
                           (unsigned int)TICKS_TO_SECS(now),
                           (unsigned int)TICKS_TO_CYCLES(now),
                           (unsigned int)TICKS_TO_OFFSET(now));
-    debugOutput(DEBUG_LEVEL_VERBOSE,"  Stop at              : %011lu (%03us %04uc %04ut)\n",
+    debugOutput(DEBUG_LEVEL_VERBOSE,"  Stop at              : %011"PRIu64" (%03us %04uc %04ut)\n",
                           tx,
                           (unsigned int)TICKS_TO_SECS(tx),
                           (unsigned int)TICKS_TO_CYCLES(tx),
@@ -1900,7 +1900,7 @@ StreamProcessor::dumpInfo()
         debugError("No handler for stream??\n");
     }
     uint64_t now = m_1394service.getCycleTimerTicks();
-    debugOutputShort( DEBUG_LEVEL_NORMAL, "  Now                   : %011lu (%03us %04uc %04ut)\n",
+    debugOutputShort( DEBUG_LEVEL_NORMAL, "  Now                   : %011"PRIu64" (%03us %04uc %04ut)\n",
                         now,
                         (unsigned int)TICKS_TO_SECS(now),
                         (unsigned int)TICKS_TO_CYCLES(now),
